fix csds parsing
